+++ ChangeLog.md
@@ -1,11 +1,17 @@
-## [0.5.21-beta.1](https://github.com/Linkontoask/v-easy-components/compare/0.5.10...0.5.21-beta.1) (2019-11-07)
+## [0.6.1-beta.2](https://github.com/Linkontoask/v-easy-components/compare/0.5.10...0.6.1-beta.2) (2019-11-13)
 
 
 ### Bug Fixes
 
+* core-js@3 to core-js@2 and useBuiltIns: "entry" ([25cc285](https://github.com/Linkontoask/v-easy-components/commit/25cc28543972c5c1e3147ba286bdfd1fc51ddbd2))
+* directive loading-preload unbind hook: "NotFoundError: Failed to execute 'removeChild' on 'Node' ([f7b0ef2](https://github.com/Linkontoask/v-easy-components/commit/f7b0ef2c5e1409f8418ddb80f78a4057b996c8fb))
+* import add Message ([9e2bdab](https://github.com/Linkontoask/v-easy-components/commit/9e2bdabc57fcf1ebc1eb78b652560dac875d4388))
+* libraryExport: 'default' config ([ac5b3ab](https://github.com/Linkontoask/v-easy-components/commit/ac5b3ab3f796914d460a312f9e55bd84606a22c5))
 * Optimize automatic deletion of tip instances and increase test code ([f228354](https://github.com/Linkontoask/v-easy-components/commit/f22835437d52e7173f5469ae9848341bd812eec4))
 * push error ([43060db](https://github.com/Linkontoask/v-easy-components/commit/43060db205733c4feb5afad3044f1ec11368f2b5))
 * use npm install prompt core-js library is not installed ([4f9f907](https://github.com/Linkontoask/v-easy-components/commit/4f9f90757d36ab63cb08f8424022f82009b59644))
+* v-image v-show bug ([82d1709](https://github.com/Linkontoask/v-easy-components/commit/82d17094e11d0074608a62cebcf6568a9b3f87ba))
+* v-loading-preload init fix and add validate-commit-msg config ([916fdae](https://github.com/Linkontoask/v-easy-components/commit/916fdae84b1b8097b99356d0bb5025c3c1805390))
 
 
 ### Features
@@ -13,7 +19,9 @@
 * Add ESLint function ([0e49e9c](https://github.com/Linkontoask/v-easy-components/commit/0e49e9cdfdb6724b7531386cbbdc06806240266e))
 * commit check ([045a59a](https://github.com/Linkontoask/v-easy-components/commit/045a59a05fd02f97228be83589bb0d7b80217b62))
 * Submit log automation and remove node-notifier ([8df5303](https://github.com/Linkontoask/v-easy-components/commit/8df5303926b3e8bf5f19783b18715fbd42c9ce89))
+* Support v-image upload value in img tag ([1eb1355](https://github.com/Linkontoask/v-easy-components/commit/1eb1355f23ecd05644b354d0ebc2d7653441388b))
 * Tool-tip Add autoRemoveTip attribute ([8131d45](https://github.com/Linkontoask/v-easy-components/commit/8131d45d09ef8dcb6041c1fb18321bf56093780d))
+* v-loading-preload support Modifier ([4233d24](https://github.com/Linkontoask/v-easy-components/commit/4233d24c62ec78d6742e98a4decf902d668420b3))